math: Pure V pow impl, rewrite ldexp and frexp to prevent infinity recursion

+++ vlib/math/pow.js.v
@@ -2,6 +2,7 @@ module math
 
 fn JS.Math.pow(x f64, y f64) f64
 
+// pow returns base raised to the provided power.
 pub fn pow(x f64, y f64) f64 {
-	return JS.Math.pow(x, y)
+	return f64(JS.Math.pow(x, y))
 }
